Implemented Redis

+++ BBSApi.Core/Models/Customer/Address.cs
@@ -2,7 +2,7 @@
 {
     public class Address
     {
-        public int AddressId { get; set; }
+        public long AddressId { get; set; }
         public string AddressLine1 { get; set; }
         public string AddressLine2 { get; set; }
         public string Town { get; set; }
